renamed osquery_thrift -> osquery, updated references
