+++ sources/osquery.js
